@@ -1,6 +1,6 @@
 {
   "private": true,
-  "version": "3.0.0-beta.1",
+  "version": "4.0.0-beta.0",
   "description": "Angular components for Google Maps",
   "repository": {
     "type": "git",
@@ -42,35 +42,34 @@
   },
   "homepage": "https://angular-maps.com",
   "dependencies": {
-    "@angular/animations": "^14.0.0",
-    "@angular/common": "^14.0.0",
-    "@angular/compiler": "^14.0.0",
-    "@angular/core": "^14.0.0",
-    "@angular/forms": "^14.0.0",
-    "@angular/platform-browser": "^14.0.0",
-    "@angular/platform-browser-dynamic": "^14.0.0",
-    "@angular/router": "^14.0.0",
-    "rxjs": "~6.6.0",
+    "@angular/animations": "^15.2.9",
+    "@angular/common": "^15.2.9",
+    "@angular/compiler": "^15.2.9",
+    "@angular/core": "^15.2.9",
+    "@angular/forms": "^15.2.9",
+    "@angular/platform-browser": "^15.2.9",
+    "@angular/platform-browser-dynamic": "^15.2.9",
+    "@angular/router": "^15.2.9",
+    "rxjs": "~7.8.0",
     "tslib": "^2.4.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.0.0",
-    "@angular/cli": "^14.0.0",
-    "@angular/compiler-cli": "^14.0.0",
-    "@compodoc/compodoc": "^1.1.10",
-    "@types/jest": "^26.0.4",
-    "@types/node": "^14.0.22",
-    "codelyzer": "^6.0.0",
-    "conventional-changelog": "^3.1.21",
-    "jest": "^26.1.0",
-    "jest-preset-angular": "^8.2.1",
-    "ng-packagr": "^14.0.0",
-    "protractor": "~7.0.0",
-    "rimraf": "^3.0.2",
-    "ts-node": "~8.10.2",
+    "@angular-devkit/build-angular": "^15.2.8",
+    "@angular/cli": "^15.2.8",
+    "@angular/compiler-cli": "^15.2.9",
+    "@compodoc/compodoc": "^1.1.20",
+    "@types/jest": "^28.0.0",
+    "@types/node": "^18.15.3",
+    "codelyzer": "^6.0.2",
+    "conventional-changelog": "^3.1.25",
+    "jest": "^28.0.0",
+    "jest-preset-angular": "^12.2.6",
+    "ng-packagr": "^15.2.2",
+    "rimraf": "^4.4.1",
+    "ts-node": "~10.9.1",
     "tslint": "~6.1.0",
-    "typescript": "~4.6.4",
-    "yarn": "^1.22.4"
+    "typescript": "~4.9.5",
+    "yarn": "^1.22.19"
   }
 }
