@@ -75,10 +75,9 @@ describe('PolygonManager', () => {
             [PolygonManager, GoogleMapsAPIWrapper],
             (polygonManager: PolygonManager, apiWrapper: GoogleMapsAPIWrapper) => {
         paths = new MvcArrayMock<google.maps.MVCArray<google.maps.LatLng>>();
-        let path = new MvcArrayMock<google.maps.LatLng>();
+        const path: MvcArrayMock<google.maps.LatLng> = new MvcArrayMock<google.maps.LatLng>();
         path.push(initLatLng);
         paths.push(path);
-
         const polygonInstance: any = {
           getPaths: () => paths,
           setMap: jest.fn(),
@@ -113,7 +112,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           paths.push(new MvcArrayMock());
           paths.push(new MvcArrayMock());
         });
@@ -140,7 +139,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           paths.pop();
           paths.removeAt(0);
         });
@@ -166,7 +165,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           const firstMvcArray = new MvcArrayMock<google.maps.LatLng>();
           firstMvcArray.push(initLatLng);
           firstMvcArray.push(initLatLng);
@@ -196,7 +195,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           paths.getAt(0).push(initLatLng);
           paths.getAt(0).push(initLatLng);
         });
@@ -223,7 +222,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           paths.getAt(0).pop();
           paths.getAt(0).removeAt(0);
         });
@@ -249,7 +248,7 @@ describe('PolygonManager', () => {
             if (expectationIndex === expectations.length) {
               done();
             }
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           paths.push(new MvcArrayMock<google.maps.LatLng>());
           paths.getAt(1).push(initLatLng);
         });
@@ -270,7 +269,7 @@ describe('PolygonManager', () => {
               eventName: 'removed_at',
               ...expectations[expectationIndex++],
             });
-          }, fail /* shouldn't have errors */, fail /* shouldn't finish */);
+          });
           const removedPath = paths.pop();
           removedPath.pop();
           done();
